Mavenize the build
