+++ README.md
@@ -1,24 +1,72 @@
-# password_vault
-A Vault to help you maintain your passwords locally
+# Password vault
 
+#### A single password to remember it all.
 
-### Required libraries:
+- A secure local vault to help you maintain your passwords.
+- Write credentials to it safely and carry the vault file everywhere. 
+- Retrieve your passwords whenever you want.
+- Strong `"AES/CBC/PKCS5Padding"` security.
 
-- lib/ 
-  - json4s/
-    - json4s-ast_2.12-3.6.0.jar
-    - json4s-core_2.12-3.6.0.jar
-    - json4s-native_2.12-3.6.0.jar
-    - json4s-scalap_2.12-3.6.0.jar
-    - paranamer-2.8.jar
-    - scala-library-2.12.6.jar
-  - commons-codec-1.15-bin.tar.gz
-  - hamcrest-core-1.3.jar
-  - json-simple-1.1.1.jar
-  - junit-4.10.jar
-  - paranamer-2.8.jar
-  - scala-library-2.12.7.jar
-  - scalactic_2.12-3.2.9.jar
-  - scalatest-app_2.12-3.2.9.jar
-  - scalatest_sjs1_2.12-3.2.9.jar
-  - sqlite-jdbc-3.20.1.jar
+**The vault file:** `~/vault_${username}.db`
+
+---
+# The software
+
+**Version:** 1.0.0
+
+**Link:** https://www.dropbox.com/s/zb51d76w60ljwnc/password_vault-1.0.0.jar?dl=0
+
+**Md5sum:** 0305ba513848ad57c12703dd149c9b65
+
+---
+# Execute
+
+Execute like a regular jar file.
+```
+ Srini: ~/password_vault (addBuild) => java -jar password_vault-1.0.0.jar 
+```
+
+---
+# Developer
+
+### Maven build
+
+Install maven on your machine using https://docs.scala-lang.org/tutorials/scala-with-maven.html
+
+To compile:
+```
+Srini: ~/password_vault (addBuild) => mvn compile
+```
+
+To execute:
+```
+Srini: ~/password_vault (addBuild) => mvn exec:java
+```
+
+To create the shippable jar:
+```
+Srini: ~/password_vault (addBuild) => mvn clean compile package assembly:single
+```
+that creates the 
+```
+ Srini: ~/password_vault (addBuild) => ls -lah target/password_vault-1.0.0.jar 
+-rw-r--r--  1 srinivasan  staff    25M Jul 11 16:47 target/password_vault-1.0.0.jar
+```
+
+For other maven commands, refer its official documentation 
+
+---
+### Set up IntelliJ
+
+Install the Maven plugin and enable `Use plugin registry` in settings.
+
+Attached the `password_vault.iml` for reference.
+
+Setup a project from existing sources using this project and then you should 
+be ale to do regular build/debug/etc.
+
+---
+### Unit tests
+
+Run tests under `VaultTest` either using Maven or via IntelliJ 
+
