+++ password_vault.iml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module org.jetbrains.idea.maven.project.MavenProjectsManager.isMavenModule="true" type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_5">
+    <output url="file://$MODULE_DIR$/target/classes" />
+    <output-test url="file://$MODULE_DIR$/target/test-classes" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/main/scala" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test/scala" isTestSource="true" />
+      <excludeFolder url="file://$MODULE_DIR$/.idea" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="jdk" jdkName="1.8" jdkType="JavaSDK" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="scala-sdk-2.12.12" level="project" />
+    <orderEntry type="library" name="Maven: org.scala-lang:scala-library:2.12.12" level="project" />
+    <orderEntry type="library" name="Maven: org.json4s:json4s-native_2.12:3.7.0-M6" level="project" />
+    <orderEntry type="library" name="Maven: org.json4s:json4s-core_2.12:3.7.0-M6" level="project" />
+    <orderEntry type="library" name="Maven: org.json4s:json4s-ast_2.12:3.7.0-M6" level="project" />
+    <orderEntry type="library" name="Maven: org.json4s:json4s-scalap_2.12:3.7.0-M6" level="project" />
+    <orderEntry type="library" name="Maven: com.thoughtworks.paranamer:paranamer:2.8" level="project" />
+    <orderEntry type="library" name="Maven: commons-codec:commons-codec:1.15" level="project" />
+    <orderEntry type="library" name="Maven: org.scalatest:scalatest-app_2.12:3.2.2" level="project" />
+    <orderEntry type="library" name="Maven: org.scala-lang:scala-reflect:2.12.12" level="project" />
+    <orderEntry type="library" name="Maven: org.xerial:sqlite-jdbc:3.20.1" level="project" />
+    <orderEntry type="library" name="Maven: junit:junit:4.12" level="project" />
+    <orderEntry type="library" name="Maven: org.hamcrest:hamcrest-core:1.3" level="project" />
+  </component>
+</module>
