+++ src/main/scala/com/pv/common/manager/VaultManager.scala
@@ -1,0 +1,27 @@
+/*
+ *  Created on: July 6, 2021
+ *      Author: Srinivasan PS
+ */
+package com.pv.common.manager
+
+object VaultManager {
+
+  def initialize(handlers: Handlers): VaultManager =
+    VaultManagerImpl.initialize(handlers = handlers)
+
+}
+
+trait VaultManager{
+
+  def addCredential()
+
+  def viewCredential()
+
+  def listAllCredentials()
+
+  def editCredential()
+
+  def deleteCredential()
+
+  def syncWithDropbox()
+}
