+++ src/main/scala/com/pv/common/manager/VaultManagerUtil.scala
@@ -1,0 +1,18 @@
+/*
+ *  Created on: July 6, 2021
+ *      Author: Srinivasan PS
+ */
+package com.pv.common.manager
+
+import com.pv.common.vault.memoized_metadata.credential.CredentialConfigs
+
+object VaultManagerUtil {
+
+  def writeCredentialToVault(
+    handlers: Handlers,
+    credentials: CredentialConfigs
+  ): Unit = {
+
+  }
+
+}
