--- src/main/util/manager/VaultManagerUtil.scala
@@ -1,18 +1,0 @@
-/*
- *  Created on: July 6, 2021
- *      Author: Srinivasan PS
- */
-package util.manager
-
-import util.vault.memoized_metadata.credential.CredentialConfigs
-
-object VaultManagerUtil {
-
-  def writeCredentialToVault(
-    handlers: Handlers,
-    credentials: CredentialConfigs
-  ): Unit = {
-
-  }
-
-}
